Move `MineBitcoinBlock` to commands/mining and remove `tenure` mod
