+++ testnet/stacks-node/src/tests/signer/commands/mod.rs
@@ -7,7 +7,6 @@ mod commit_ops;
 mod mining;
 mod shutdown;
 mod stalling;
-mod tenure;
 mod transfer;
 
 pub use context::{SignerTestContext};
@@ -18,8 +17,7 @@ pub use block_commit::{
 pub use block_wait::{WaitForBlockFromMiner1, WaitForBlockFromMiner2};
 pub use boot::BootToEpoch3;
 pub use commit_ops::{SkipCommitOpPrimaryMiner, SkipCommitOpSecondaryMiner};
-pub use mining::MineBitcoinBlockTenureChangePrimaryMiner;
+pub use mining::{MineBitcoinBlock, MineBitcoinBlockTenureChangePrimaryMiner};
 pub use shutdown::ShutdownMiners;
 pub use stalling::{RecoverFromStall, StallMining};
-pub use tenure::MineBitcoinBlock;
 pub use transfer::SendTransferTx;
